chore: move TopNavigationBar to search folder
