+++ app/(tabs)/search.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { SafeAreaView } from 'react-native';
 import ToggleNavigationBar from 'components/ToggleNavigationBar';
-import TopNavigationBar from 'components/TopNavigationBar';
+import TopNavigationBar from 'components/Search/TopNavigationBar';
 import CreateGatheringBtn from 'components/CreateGatheringBtn';
 import { GatheringType } from 'types/Gatherings';
