+++ components/Search/TopNavigationBar.tsx
@@ -1,6 +1,6 @@
 import { Text } from 'react-native';
 import { createMaterialTopTabNavigator } from '@react-navigation/material-top-tabs';
-import GatheringList from './Search/GatheringList';
+import GatheringList from './GatheringList';
 import { tokens }  from 'constants/';
 import { GatheringType } from 'types/Gatherings';
